Split BackingStore into BackingStore and Backend
BackingStore takes the in-memory window structs, serializes them using
Cap'n Proto and calls the Backend. The key is still the pair of
streamID, windowID. Does the same for "Get"s as well.

The Backend is an interface that operates on (streamID, windowID, buffer),
storing and retrieving them as needed. Currently a simple map-based
InMemoryBackend is implemented.