Add badger backend
Currently, both landmark and summary windows are stored in the same KV
database, with the following 17-byte key:

   0          1           9           17
    <LANDMARK> <WINDOW ID> <STREAM ID>

Currently, we don't handle an txn errors. We kill the program if we face
an error. We would want to do retries in the same of failures. The
failure case will be *super* rare since that will requires two
messages with the same timestamp to the same stream.