Fix data races
There were two cases of data races:

1. All the sentinels used for shutdown/flush were accessed in different
goroutines without any locking. The race is because the sentinels are
written to be static "singleton"s. The fix is to guard the function
call for getting the sentinel with a mutex.

2. For testing SummaryDB, there's a simple map-based backend called
"InMemoryBackend". However, the maps were not guarded. Use a mutex
while accessing.

GitHub action for testing now also tests for race-conditions using
the `-race` flag.