Parallelize issuing pending merges using a work group
The number of goroutines spawned isn't an issue since the number
of pending merges is likely to be in the order of 10^2. If this
ever becomes an issue, a worker pool can be used to control
the number of goroutines.

This makes the pipeline benchmark faster (5-10%), especially
when merge window sizes are small, with a large pending merges
buffers.