[Storage] Add support to persist MinHeap on disk
Uses the KV-store backend to store the heap on disk, with the
key being the stream ID. The MinHeap is serialized using gob
with performance numbers:
	Serialize - 140ns/item
	Deserialize - 350ns/item, dropping to 170ns/item (between
		      100 and 1000 heap items)

gob is used instead of CapnProto (like the other structs) due to
convienience. There's a TODO to implement serialization in CapnProto
and compare the performance.