[Pipeline/Merger] Flush heap to disk when flusing pending merges
Extend StreamWindowManager to allow I/O of heaps through the
BackingStore, simialr to windows.
TODO: Rename this to "StorageManager" since it is beyond the storing
just storing windows. In the future, this may be used for WAL too.

Fix small bug where the time-stamp for storing a heap in Badger was
taken as 0, and use the same key structure as for windows. This
caused a problem when an actual window at t=0 was to be stored.
For heaps, take the timestamp to now be MinInt64.