[Pipeline] Add tests for summarizer and ingestor
The primary reason for this change was to catch and fix
a bug in ingestor. The append which caused a buffer swap
(current buffer is full, allocate new) would be lost.

Move the IngestBufferAllocator to a separate file, and
introduce an allocator interface for testing allocations
and deallocations.

SummarizerTest is a new type that makes writing tests
easier. The coverage of summarizer test is not extensive.
It currently covers only the base cases: single full buffer,
single full buffer + single partial buffer, multiple full
buffers.

Closes #3